refactor and separate player actions tree from framerules
